Implement FSSpecFSDPStrategy
