+++ dataflux_pytorch/benchmark/checkpointing/multinode/train.py
@@ -6,6 +6,7 @@ import torch
 import statistics
 import argparse
 
+from fsspecfsdp import FSSpecFSDPStrategy
 from lightning import Trainer
 from lightning.pytorch.strategies import FSDPStrategy
 from lightning.pytorch.callbacks import ModelCheckpoint
@@ -58,7 +59,9 @@ def main(project: str,
             state_dict_type="sharded",
         )
     else:
-        strategy = FSDPStrategy(state_dict_type="sharded")
+        strategy = FSSpecFSDPStrategy(path=ckpt_dir_path,
+                                      model=model,
+                                      state_dict_type="sharded")
     min_epochs_save = int(os.environ.get("MIN_EPOCHS_SAVE", 4))
     max_epochs_save = int(os.environ.get("MAX_EPOCHS_SAVE", 5))
     max_steps_save = int(os.environ.get("MAX_STEPS_SAVE", 3))
@@ -111,7 +114,9 @@ def main(project: str,
                 state_dict_type="sharded",
             )
         else:
-            strategy = FSDPStrategy(state_dict_type="sharded")
+            strategy = FSSpecFSDPStrategy(path=ckpt_dir_path,
+                                          model=model,
+                                          state_dict_type="sharded")
         trainer = Trainer(
             default_root_dir=ckpt_dir_path,
             plugins=[],
